+++ VerteilteSystemeProjekt/src/main/java/de/vsys/project/general/data/IncidenceValueLastSevenDays.java
@@ -5,9 +5,9 @@ public class IncidenceValueLastSevenDays {
     @author: David Rohrschneider
     This class is the template for the data which we return to the user
      */
-    private int incidenceValueLastSevenDays;
+    private double incidenceValueLastSevenDays;
 
-    public int getIncidenceValueLastSevenDays() {
+    public double getIncidenceValueLastSevenDays() {
         return incidenceValueLastSevenDays;
     }
